requirements_dev.txt
bumpversion==0.5.3
flake8==3.5.0
ipdb==0.12
tox==3.5.2
coverage==4.5.1
python-dateutil==2.8.0
Sphinx==1.8.1
twine==1.12.1
watchdog==0.9.0
wheel==0.32.1

pytest==4.6.3
pytest-runner==4.2
